Crate javascriptcore

JavaScriptCore Bindings

Evaluate JavaScript programs from within an app, and support JavaScript scripting of your app.

Structs

JSClass
A JavaScript class.

JSContext
A JavaScript execution context.

JSContextGroup
A group that associates JavaScript contexts with one another.

JSObject
A JavaScript object.

JSString
A UTF16 character buffer.

JSValue
A JavaScript value.

Enums

JSType

JSTypedArrayType

Functions

check_script_syntax
Checks for syntax errors in a string of JavaScript.

evaluate_script
Evaluates a string of JavaScript.

garbage_collect
Performs a JavaScript garbage collection.