Struct javascriptcore::JSContext

pub struct JSContext { /* fields omitted */ }

A JavaScript execution context.

Holds the global object and other execution state.

Methods

impl JSContext

fn new() -> Self

Creates a global JavaScript execution context and populates it with all the built-in JavaScript objects, such as Object, Function, String, and Array.

In WebKit version 4.0 and later, the context is created in a unique context group. Therefore, scripts may execute in it concurrently with scripts executing in other contexts. However, you may not use values created in the context in other contexts.

fn new_with_class(global_object_class: &JSClass) -> Self

Creates a global JavaScript execution context and populates it with all the built-in JavaScript objects, such as Object, Function, String, and Array.

In WebKit version 4.0 and later, the context is created in a unique context group. Therefore, scripts may execute in it concurrently with scripts executing in other contexts. However, you may not use values created in the context in other contexts.

global_object_class: The class to use when creating the global object.