Struct javascriptcore::JSObject

pub struct JSObject { /* fields omitted */ }

A JavaScript object.

An JSObject is a JSValue. This is implemented by having JSObject implement the Deref trait so that anything that expects a JSValue can receive a JSObject as well.

Methods

impl JSObject

fn property_names(&self) -> JSObjectPropertyNameIter

Gets an iterator over the names of an object's enumerable properties.

let ctx = JSContext::default();
let v = JSValue::new_from_json(&ctx, "{\"id\": 123}").expect("valid object");
let o = v.as_object().expect("object");

let names: Vec<String> = o.property_names()
                          .map(|s| s.to_string())
                          .collect();
assert_eq!(names, vec!["id"]);

fn has_property<S>(&self, name: S) -> bool where
S: Into<JSString>,

Tests whether an object has a given property.

name: A value that can be converted to a JSString containing the property's name.

Returns true if the object has a property whose name matches name, otherwise false.

let ctx = JSContext::default();
let v = JSValue::new_from_json(&ctx, "{\"id\": 123}").expect("valid object");
let o = v.as_object().expect("object");

assert!(o.has_property("id"));

fn get_property<S>(&self, name: S) -> JSValue where
S: Into<JSString>,

Gets a property from an object.

name: A value that can be converted to a JSString containing the property's name.

Returns the property's value if object has the property, otherwise the undefined value.

let ctx = JSContext::default();
let v = JSValue::new_from_json(&ctx, "{\"id\": 123}").expect("valid object");
let o = v.as_object().expect("object");

let n = o.get_property("id");
assert!(n.is_number());
// Remember that this will be an f64 now!
assert_eq!(n.as_number().expect("number"), 123.0);

fn get_property_at_index(&self, index: u32) -> JSValue

Gets a property from an object by numeric index.

index: An integer value that is the property's name.

Returns the property's value if object has the property, otherwise the undefined value.

Calling get_property_at_index is equivalent to calling get_property with a string containing index, but get_property_at_index provides optimized access to numeric properties.

let ctx = JSContext::default();
let v = JSValue::new_from_json(&ctx, "[3, true, \"abc\"]").expect("valid array");
let o = v.as_object().expect("object");

let n = o.get_property_at_index(0).as_number().expect("number");
let b = o.get_property_at_index(1).as_boolean();
let s = o.get_property_at_index(2).as_string().expect("string");

assert_eq!(n, 3.0);
assert_eq!(b, true);
assert_eq!(s, "abc".into());