Struct javascriptcore::JSValue

pub struct JSValue { /* fields omitted */ }

A JavaScript value.

The base type for all JavaScript values, and polymorphic functions on them.

Methods

impl JSValue

fn new_undefined(ctx: &JSContext) -> Self

Creates a JavaScript value of the undefined type.

ctx: The execution context to use.

Returns the unique undefined value.

fn new_null(ctx: &JSContext) -> Self

Creates a JavaScript value of the null type.

ctx: The execution context to use.

Returns the unique null value.

fn new_boolean(ctx: &JSContext, boolean: bool) -> Self

Creates a JavaScript value of the boolean type.

ctx: The execution context to use.

boolean: The bool to assign to the newly created JSValue.

Returns a JSValue of the boolean type, representing the value of boolean.

fn new_number(ctx: &JSContext, number: f64) -> Self

Creates a JavaScript value of the number type.

ctx: The execution context to use.

number: The f64 to assign to the newly created JSValue.

Returns a JSValue of the number type, representing the value of number.

fn new_string<S: Into<JSString>>(ctx: &JSContext, string: S) -> Self

Creates a JavaScript value of the string type.

ctx: The execution context to use.

string: The JSString to assign to the newly created JSValue. The newly createdJSValue` retains string, and releases it upon garbage collection.

Returns a JSValue of the string type, representing the value of string.

fn new_from_json<S: Into<JSString>>(
ctx: &JSContext,
string: S
) -> Result<Self, ()>

Creates a JavaScript value from a JSON formatted string.

ctx: The execution context to use.

string: A value that can be converted into a JSString containing the JSON string to be parsed.

Returns a Result with the JSValue containing the parsed value, or an error if the input is invalid.

fn to_json_string(&self, indent: u32) -> Result<JSString, JSException>

Creates a JavaScript string containing the JSON serialized representation of a JS value.

indent: The number of spaces to indent when nesting. If 0, the resulting JSON will not contains newlines. The size of the indent is clamped to 10 spaces.

Returns either a JSString with the result of serialization, or an exception if one was thrown.

fn get_type(&self) -> JSType

Returns a JavaScript value's type.

Returns a value of type JSType that identifies value's type.

fn is_undefined(&self) -> bool

Tests whether a JavaScript value's type is the undefined type.

Returns true if value's type is the undefined type, otherwise false.

fn is_null(&self) -> bool

Tests whether a JavaScript value's type is the null type.

Returns true if value's type is the null type, otherwise false.

fn is_boolean(&self) -> bool

Tests whether a JavaScript value's type is the boolean type.

Returns true if value's type is the boolean type, otherwise false.

fn is_number(&self) -> bool

Tests whether a JavaScript value's type is the number type.

Returns true if value's type is the number type, otherwise false.

fn is_string(&self) -> bool

Tests whether a JavaScript value's type is the string type.

Returns true if value's type is the string type, otherwise false.

fn is_object(&self) -> bool

Tests whether a JavaScript value's type is the object type.

Returns true if value's type is the object type, otherwise false.

fn is_object_of_class(&self, js_class: &JSClass) -> bool

Tests whether a JavaScript value is an object with a given class in its class chain.

js_class: The JSClass to test against.

Returns true if value is an object and has jsClass in its class chain, otherwise false.

fn is_array(&self) -> bool

Tests whether a JavaScript value is an array.

Returns true if value is an array, otherwise false.

fn is_date(&self) -> bool

Tests whether a JavaScript value is a date.

Returns true if value is a date, otherwise false.

fn as_boolean(&self) -> bool

Converts a JavaScript value to boolean and returns the resulting boolean.

Returns the boolean result of conversion.

fn as_number(&self) -> Result<f64, JSException>

Converts a JavaScript value to number and returns the resulting number.

Returns either the numeric result of conversion, or an exception if one was thrown.

fn as_string(&self) -> Result<JSString, JSException>

Converts a JavaScript value to string and copies the result into a JavaScript string.

Returns either JSString with the result of conversion, or an exception if one was thrown. Ownership follows the Create Rule.

fn as_object(&self) -> Result<JSObject, JSException>

Converts a JavaScript value to object and returns the resulting object.

Returns either the JSObject result of conversion, or an exception if one was thrown.