Struct javascriptcore::JSObject

pub struct JSObject { /* fields omitted */ }

A JavaScript object.

An JSObject is a JSValue. This is implemented by having JSObject implement the Deref trait so that anything that expects a JSValue can receive a JSObject as well.

Methods

impl JSObject

fn property_names(&self) -> JSObjectPropertyNameIter

Gets an iterator over the names of an object's enumerable properties.

let names: Vec<JSString> = obj.property_names().collect();

fn has_property<S>(&self, name: S) -> bool where
S: Into<JSString>,

Tests whether an object has a given property.

name: A value that can be converted to a JSString containing the property's name.

Returns true if the object has a property whose name matches name, otherwise false.

if obj.has_property("id") {
    // ...
}

fn get_property<S>(&self, name: S) -> JSValue where
S: Into<JSString>,

Gets a property from an object.

name: A value that can be converted to a JSString containing the property's name.

Returns the property's value if object has the property, otherwise the undefined value.

fn get_property_at_index(&self, index: u32) -> JSValue

Gets a property from an object by numeric index.

index: An integer value that is the property's name.

Returns the property's value if object has the property, otherwise the undefined value.

Calling get_property_at_index is equivalent to calling get_property with a string containing index, but get_property_at_index provides optimized access to numeric properties.