Struct javascriptcore::JSValue

pub struct JSValue { /* fields omitted */ }

A JavaScript value.

The base type for all JavaScript values, and polymorphic functions on them.

All values passed between Rust and JavaScriptCore will be boxed with a JSValue.

Creating JS values

JSValue::new_undefined

JSValue::new_null

JSValue::new_boolean

JSValue::new_number

JSValue::new_string

JSValue::new_from_json

JSON

JSValue::new_from_json

JSValue::to_json_string

Retrieving Rust values

JSValue::as_boolean

JSValue::as_number

JSValue::as_object

JSValue::as_string

Methods

impl JSValue

fn new_undefined(ctx: &JSContext) -> Self

Creates a JavaScript value of the undefined type.

ctx: The execution context to use.

Returns the unique undefined value.

let ctx = JSContext::default();

let v = JSValue::new_undefined(&ctx);
assert!(v.is_undefined());

fn new_null(ctx: &JSContext) -> Self

Creates a JavaScript value of the null type.

ctx: The execution context to use.

Returns the unique null value.

let ctx = JSContext::default();

let v = JSValue::new_null(&ctx);
assert!(v.is_null());

fn new_boolean(ctx: &JSContext, boolean: bool) -> Self

Creates a JavaScript value of the boolean type.

ctx: The execution context to use.

boolean: The bool to assign to the newly created JSValue.

Returns a JSValue of the boolean type, representing the value of boolean.

let ctx = JSContext::default();

let v = JSValue::new_boolean(&ctx, true /* or false */);
assert!(v.is_boolean());

fn new_number(ctx: &JSContext, number: f64) -> Self

Creates a JavaScript value of the number type.

ctx: The execution context to use.

number: The f64 to assign to the newly created JSValue.

Returns a JSValue of the number type, representing the value of number.

let ctx = JSContext::default();

let v = JSValue::new_number(&ctx, 3.0f64);
assert!(v.is_number());

let v = JSValue::new_number(&ctx, 3.0f32 as f64);
assert!(v.is_number());

let v = JSValue::new_number(&ctx, 3 as f64);
assert!(v.is_number());

fn new_string<S: Into<JSString>>(ctx: &JSContext, string: S) -> Self

Creates a JavaScript value of the string type.

ctx: The execution context to use.

string: The JSString to assign to the newly created JSValue. The newly created JSValue retains string, and releases it upon garbage collection.

Returns a JSValue of the string type, representing the value of string.

let ctx = JSContext::default();

let v = JSValue::new_string(&ctx, "abc");
assert!(v.is_string());

fn new_from_json<S: Into<JSString>>(
ctx: &JSContext,
string: S
) -> Result<Self, ()>

Creates a JavaScript value from a JSON formatted string.

ctx: The execution context to use.

string: A value that can be converted into a JSString containing the JSON string to be parsed.

Returns a Result with the JSValue containing the parsed value, or an error if the input is invalid.

let ctx = JSContext::default();

let v = JSValue::new_from_json(&ctx, "true").expect("value");
assert!(v.is_boolean());

fn to_json_string(&self, indent: u32) -> Result<JSString, JSException>

Creates a JavaScript string containing the JSON serialized representation of a JS value.

indent: The number of spaces to indent when nesting. If 0, the resulting JSON will not contains newlines. The size of the indent is clamped to 10 spaces.

Returns either a JSString with the result of serialization, or an exception if one was thrown.

let ctx = JSContext::default();

let v = JSValue::new_boolean(&ctx, false);
let s = v.to_json_string(0).unwrap();
// The result is a `JSString`, so we jump through a hoop here.
assert_eq!(s, "false".into());

fn get_type(&self) -> JSType

Returns a JavaScript value's type.

Returns a value of type JSType that identifies value's type.

let ctx = JSContext::default();

let v = JSValue::new_from_json(&ctx, "true").expect("value");
assert_eq!(v.get_type(), JSType::Boolean);

let v = JSValue::new_from_json(&ctx, "5.0").expect("value");
assert_eq!(v.get_type(), JSType::Number);

let v = JSValue::new_from_json(&ctx, "null").expect("value");
assert_eq!(v.get_type(), JSType::Null);

fn is_undefined(&self) -> bool

Tests whether a JavaScript value's type is the undefined type.

Returns true if value's type is the undefined type, otherwise false.

let ctx = JSContext::default();

let v = JSValue::new_undefined(&ctx);
assert!(v.is_undefined());

fn is_null(&self) -> bool

Tests whether a JavaScript value's type is the null type.

Returns true if value's type is the null type, otherwise false.

let ctx = JSContext::default();

let v = JSValue::new_from_json(&ctx, "null").expect("value");
assert!(v.is_null());

fn is_boolean(&self) -> bool

Tests whether a JavaScript value's type is the boolean type.

Returns true if value's type is the boolean type, otherwise false.

let ctx = JSContext::default();

let v = JSValue::new_from_json(&ctx, "false").expect("value");
assert!(v.is_boolean());

fn is_number(&self) -> bool

Tests whether a JavaScript value's type is the number type.

Returns true if value's type is the number type, otherwise false.

let ctx = JSContext::default();

let v = JSValue::new_from_json(&ctx, "-23").expect("value");
assert!(v.is_number());

fn is_string(&self) -> bool

Tests whether a JavaScript value's type is the string type.

Returns true if value's type is the string type, otherwise false.

let ctx = JSContext::default();

let v = JSValue::new_from_json(&ctx, "\"Pueri et puellae\"").expect("value");
assert!(v.is_string());

fn is_object(&self) -> bool

Tests whether a JavaScript value's type is the object type.

Returns true if value's type is the object type, otherwise false.

let ctx = JSContext::default();

let v = JSValue::new_from_json(&ctx, "{\"id\": 123}").expect("valid object");
assert!(v.is_object());

fn is_object_of_class(&self, js_class: &JSClass) -> bool

Tests whether a JavaScript value is an object with a given class in its class chain.

js_class: The JSClass to test against.

Returns true if value is an object and has jsClass in its class chain, otherwise false.

fn is_array(&self) -> bool

Tests whether a JavaScript value is an array.

Returns true if value is an array, otherwise false.

fn is_date(&self) -> bool

Tests whether a JavaScript value is a date.

Returns true if value is a date, otherwise false.

fn as_boolean(&self) -> bool

Converts a JavaScript value to boolean and returns the resulting boolean.

Returns the boolean result of conversion.

let ctx = JSContext::default();

let v = JSValue::new_boolean(&ctx, false);
assert_eq!(v.as_boolean(), false);

fn as_number(&self) -> Result<f64, JSException>

Converts a JavaScript value to number and returns the resulting number.

Returns either the numeric result of conversion, or an exception if one was thrown.

let ctx = JSContext::default();

let v = JSValue::new_number(&ctx, 5.0);
let n = v.as_number().expect("valid number");
assert_eq!(n, 5.0);

fn as_string(&self) -> Result<JSString, JSException>

Converts a JavaScript value to string and copies the result into a JavaScript string.

Returns either JSString with the result of conversion, or an exception if one was thrown. Ownership follows the Create Rule.

let ctx = JSContext::default();

let v = JSValue::new_string(&ctx, "Cave canem.");
let s = v.as_string().expect("valid string");
assert_eq!(s, "Cave canem.".into());

fn as_object(&self) -> Result<JSObject, JSException>

Converts a JavaScript value to object and returns the resulting object.

Returns either the JSObject result of conversion, or an exception if one was thrown.

let ctx = JSContext::default();

let v = JSValue::new_from_json(&ctx, "{\"id\": 123}").expect("valid object");
let o = v.as_object().expect("object");
// We now have an object that we can inspect.